requirements.txt
alembic==1.13.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cymem==2.0.8
exceptiongroup==1.1.2
Flask==3.0.0
Flask-CKEditor==0.5.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==3.0.3
gspread==5.10.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.2
preshed==2.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spotipy==2.23.0
SQLAlchemy==2.0.24
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.9.0
urllib3==1.26.16
vboxapi==1.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
zipp==3.17.0
